# Conflicts:
#	src/main/java/in/koreatech/koin/global/domain/notification/repository/NotificationSubscribeRepository.java

+++ src/main/java/in/koreatech/koin/domain/community/repository/ArticleViewLogRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.community.model.ArticleViewLog;
 
-public interface ArticleViewLogRepository extends Repository<ArticleViewLog, Long> {
+public interface ArticleViewLogRepository extends Repository<ArticleViewLog, Integer> {
 
-    Optional<ArticleViewLog> findByArticleIdAndUserId(Long articleId, Long userId);
+    Optional<ArticleViewLog> findByArticleIdAndUserId(Integer articleId, Integer userId);
 
     ArticleViewLog save(ArticleViewLog articleViewLog);
 }
