+++ src/main/java/in/koreatech/koin/domain/timetable/repository/LectureRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.Repository;
 
 import in.koreatech.koin.domain.timetable.model.Lecture;
 
-public interface LectureRepository extends Repository<Lecture, Long> {
+public interface LectureRepository extends Repository<Lecture, Integer> {
 
     List<Lecture> findBySemester(String semesterDate);
