+++ src/main/java/in/koreatech/koin/domain/user/model/UserToken.java
@@ -15,20 +15,20 @@ public class UserToken {
     private static final long REFRESH_TOKEN_EXPIRE_DAY = 14L;
 
     @Id
-    private Long id;
+    private Integer id;
 
     private final String refreshToken;
 
     @TimeToLive(unit = TimeUnit.DAYS)
     private final Long expiration;
 
-    private UserToken(Long id, String refreshToken, Long expiration) {
+    private UserToken(Integer id, String refreshToken, Long expiration) {
         this.id = id;
         this.refreshToken = refreshToken;
         this.expiration = expiration;
     }
 
-    public static UserToken create(Long userId, String refreshToken) {
+    public static UserToken create(Integer userId, String refreshToken) {
         return new UserToken(userId, refreshToken, REFRESH_TOKEN_EXPIRE_DAY);
     }
 }
