+++ src/main/java/in/koreatech/koin/global/domain/notification/repository/NotificationSubscribeRepository.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 
 import org.springframework.data.repository.Repository;
 
+import in.koreatech.koin.global.domain.notification.exception.NotificationSubscribeNotFoundException;
 import in.koreatech.koin.global.domain.notification.model.NotificationSubscribe;
 import in.koreatech.koin.global.domain.notification.model.NotificationSubscribeType;
-import in.koreatech.koin.global.domain.notification.exception.NotificationSubscribeNotFoundException;
 
 public interface NotificationSubscribeRepository extends Repository<NotificationSubscribe, Long> {
 
@@ -15,16 +15,16 @@ public interface NotificationSubscribeRepository extends Repository<Notification
 
     List<NotificationSubscribe> findAllBySubscribeType(NotificationSubscribeType type);
 
-    Optional<NotificationSubscribe> findByUserIdAndSubscribeType(Long userId, NotificationSubscribeType type);
+    Optional<NotificationSubscribe> findByUserIdAndSubscribeType(Integer userId, NotificationSubscribeType type);
 
-    default NotificationSubscribe getByUserIdAndSubscribeType(Long userId, NotificationSubscribeType type) {
+    default NotificationSubscribe getByUserIdAndSubscribeType(Integer userId, NotificationSubscribeType type) {
         return findByUserIdAndSubscribeType(userId, type)
             .orElseThrow(
                 () -> NotificationSubscribeNotFoundException.withDetail("userId: " + userId + ", type: " + type)
             );
     }
 
-    void deleteByUserIdAndSubscribeType(Long userId, NotificationSubscribeType type);
+    void deleteByUserIdAndSubscribeType(Integer userId, NotificationSubscribeType type);
 
-    List<NotificationSubscribe> findAllByUserId(Long userId);
+    List<NotificationSubscribe> findAllByUserId(Integer userId);
 }
